+++ src/com/rian/difficultycalculator/utils/StandardHitWindowConverter.java
@@ -14,10 +14,9 @@ public final class StandardHitWindowConverter {
      * @param value The value of the hit window, in milliseconds.
      * @return The overall difficulty value.
      */
-    public static double hitWindow300ToOD(double value) {
+    public static float hitWindow300ToOD(float value) {
         return (80 - value) / 6;
     }
-
 
     /**
      * Calculates the overall difficulty value of an ok hit window.
@@ -25,7 +24,7 @@ public final class StandardHitWindowConverter {
      * @param value The value of the hit window, in milliseconds.
      * @return The overall difficulty value.
      */
-    public static double hitWindow100ToOD(double value) {
+    public static float hitWindow100ToOD(float value) {
         return (140 - value) / 8;
     }
 
@@ -35,7 +34,7 @@ public final class StandardHitWindowConverter {
      * @param value The value of the hit window, in milliseconds.
      * @return The overall difficulty value.
      */
-    public static double hitWindow50ToOD(double value) {
+    public static float hitWindow50ToOD(float value) {
         return (200 - value) / 10;
     }
 
@@ -45,7 +44,7 @@ public final class StandardHitWindowConverter {
      * @param od The overall difficulty value.
      * @return The hit window in milliseconds.
      */
-    public static double odToHitWindow300(double od) {
+    public static float odToHitWindow300(float od) {
         return 80 - 6 * od;
     }
 
@@ -55,7 +54,7 @@ public final class StandardHitWindowConverter {
      * @param od The overall difficulty value.
      * @return The hit window in milliseconds.
      */
-    public static double odToHitWindow100(double od) {
+    public static float odToHitWindow100(float od) {
         return 140 - 8 * od;
     }
 
@@ -65,7 +64,7 @@ public final class StandardHitWindowConverter {
      * @param od The overall difficulty value.
      * @return The hit window in milliseconds.
      */
-    public static double odToHitWindow50(double od) {
+    public static float odToHitWindow50(float od) {
         return 200 - 10 * od;
     }
 }
